+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/MonitorFactory.kt
@@ -7,27 +7,14 @@ import java.util.*
 object MonitorFactory {
     fun createMonitor(type: MonitorType, id: UUID, config: MonitorConfig): Monitor {
         return when (type) {
-            MonitorType.CPU_LOAD -> CpuMonitor(
-                id,
-                config
-            )
-            MonitorType.CPU_TEMP -> CpuTemperatureMonitor(
-                id,
-                config
-            )
-            MonitorType.DRIVE_SPACE -> DriveMonitor(
-                id,
-                config
-            )
-            MonitorType.MEMORY_SPACE -> MemoryMonitor(
-                id,
-                config
-            )
-            MonitorType.NETWORK_UP -> NetworkUpMonitor(
-                id,
-                config
-            )
+            MonitorType.CPU_LOAD -> CpuMonitor(id, config)
+            MonitorType.CPU_TEMP -> CpuTemperatureMonitor(id, config)
+            MonitorType.DRIVE_SPACE -> DriveMonitor(id, config)
+            MonitorType.MEMORY_SPACE -> MemoryMonitor(id, config)
+            MonitorType.NETWORK_UP -> NetworkUpMonitor(id, config)
             MonitorType.CONTAINER_RUNNING -> DockerContainerRunningMonitor(id, config)
+            MonitorType.PROCESS_MEMORY_SPACE -> ProcessMemoryMonitor(id, config)
+            MonitorType.PROCESS_CPU_LOAD -> ProcessCpuMonitor(id, config)
         }
     }
 }