+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/MonitorStore.kt
@@ -27,7 +27,7 @@ class MonitorStore(objectMapper: ObjectMapper) :
         )
 
         enum class Type {
-            CPU_LOAD, CPU_TEMP, DRIVE_SPACE, MEMORY_SPACE, NETWORK_UP, CONTAINER_RUNNING
+            CPU_LOAD, CPU_TEMP, DRIVE_SPACE, MEMORY_SPACE, NETWORK_UP, CONTAINER_RUNNING, PROCESS_MEMORY_SPACE, PROCESS_CPU_LOAD
         }
 
     }